clean/ クラス定義を別ファイルに切り出し
