+++ config/routes.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'haml'
+require 'Accessor'
 
 set :haml, format: :html5,
            views: './lib/views'
@@ -64,91 +65,6 @@ helpers do
       @validation.load_content
     else
       redirect '/'
-    end
-  end
-end
-
-module Memo
-  class Content
-    attr_accessor :title, :text, :text_array
-
-    def initialize(title = '')
-      return if title.empty?
-
-      @path = "./data/#{title}"
-      @title = title
-      @text = load_text
-      @text_array = convert_text_to_array
-    end
-
-    def load_text
-      File.read(@path)
-    end
-
-    def convert_text_to_array
-      @text.split(/[\n|\r\n|\r]/)
-    end
-
-    def create(title:, text:)
-      @title ||= title
-      @text = text
-      save
-    end
-
-    def update(title:, text:)
-      File.rename(@path, "./data/#{title}") if @title != title
-      @title = title
-      @text = text
-      save
-    end
-
-    def save
-      File.open(@path, 'w') { |f| f.print @text }
-    end
-
-    def destroy
-      File.delete(@path)
-    end
-  end
-
-  class Accessor
-    attr_reader :title, :path
-
-    def initialize(request_path)
-      @title = request_path.slice(/[\w-]+/)
-      @path = "./data/#{@title}"
-    end
-
-    def exit?(path = @path)
-      !Dir.glob(path).empty?
-    end
-
-    def validate_new_content(text:)
-      return false if exit?(@path)
-
-      @text = text
-      self
-    end
-
-    def validate_update_content(memo, new_content)
-      new_title = new_content[:title]
-      return false if memo.title != new_title && exit?("./data/#{new_title}")
-
-      @title = new_title
-      @text = new_content[:text]
-      self
-    end
-
-    def create_content
-      Content.new.create(title: @title, text: @text)
-    end
-
-    def load_content
-      Content.new(@title)
-    end
-
-    def update_content(old_title)
-      Content.new(old_title).update(title: @title, text: @text)
     end
   end
 end
